Fixed recalculate normals function, added a few model examples and renamed some folders
